# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Capacity & Spend Control API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /budgets:
    get:
      operationId: listBudgets
      tags: [budgets]
      parameters:
        - { name: projectId, in: query, schema: { type: integer } }
        - { name: year, in: query, schema: { type: integer } }
        - { name: fy, in: query, schema: { type: integer }, description: "Fiscal year (Jul–Jun, named by ending year)" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/BudgetPlan" }
    post:
      operationId: createBudget
      tags: [budgets]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BudgetPlanInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BudgetPlan" }
  /budgets/{id}:
    patch:
      operationId: updateBudget
      tags: [budgets]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BudgetPlanInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BudgetPlan" }
    delete:
      operationId: deleteBudget
      tags: [budgets]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Top-level KPIs across portfolio
      parameters:
        - { name: year, in: query, schema: { type: integer } }
        - { name: fy, in: query, schema: { type: integer }, description: "Fiscal year (Jul–Jun, named by ending year)" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DashboardSummary" }
  /dashboard/budget-vs-actual:
    get:
      operationId: getBudgetVsActual
      tags: [dashboard]
      summary: Approved vs forecast vs actual by month
      parameters:
        - { name: year, in: query, schema: { type: integer } }
        - { name: fy, in: query, schema: { type: integer }, description: "Fiscal year (Jul–Jun, named by ending year)" }
        - { name: projectId, in: query, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/BudgetVsActualPoint" }
  /dashboard/utilization:
    get:
      operationId: getUtilization
      tags: [dashboard]
      summary: Resource utilization by month
      parameters:
        - { name: year, in: query, schema: { type: integer } }
        - { name: fy, in: query, schema: { type: integer }, description: "Fiscal year (Jul–Jun, named by ending year)" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/UtilizationRow" }
...
  /dashboard/demand-vs-capacity:
    get:
      operationId: getDemandVsCapacity
      tags: [dashboard]
      summary: Demand vs available capacity by skill
      parameters:
        - { name: year, in: query, schema: { type: integer } }
        - { name: fy, in: query, schema: { type: integer }, description: "Fiscal year (Jul–Jun, named by ending year)" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/DemandVsCapacityRow" }
...
  /dashboard/opex-capex:
    get:
      operationId: getOpexCapex
      tags: [dashboard]
      summary: Opex vs Capex split by month
      parameters:
        - { name: year, in: query, schema: { type: integer } }
        - { name: fy, in: query, schema: { type: integer }, description: "Fiscal year (Jul–Jun, named by ending year)" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/OpexCapexRow" }
  /dashboard/skill-shortage:
    get:
      operationId: getSkillShortage
      tags: [dashboard]
      summary: Forecasted skill shortages by month
      parameters:
        - { name: year, in: query, schema: { type: integer } }
        - { name: fy, in: query, schema: { type: integer }, description: "Fiscal year (Jul–Jun, named by ending year)" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SkillShortageRow" }
components:
  schemas:
...
    BudgetPlan:
      type: object
      required:
        - id
        - projectId
        - projectName
        - year
        - month
        - costType
        - approvedBudget
        - forecastBudget
        - plannedResourceCost
        - nonResourceCost
      properties:
        id: { type: integer }
        projectId: { type: integer }
        projectName: { type: string }
        year: { type: integer }
        month: { type: integer }
        costType: { type: string, enum: [Opex, Capex] }
        approvedBudget: { type: number }
        forecastBudget: { type: number }
        plannedResourceCost: { type: number }
        nonResourceCost: { type: number }
    BudgetPlanInput:
      type: object
      required:
        - projectId
        - year
        - month
        - costType
        - approvedBudget
        - forecastBudget
        - plannedResourceCost
        - nonResourceCost
      properties:
        projectId: { type: integer }
        year: { type: integer }
        month: { type: integer }
        costType: { type: string, enum: [Opex, Capex] }
        approvedBudget: { type: number }
        forecastBudget: { type: number }
        plannedResourceCost: { type: number }
        nonResourceCost: { type: number }
...
    DashboardSummary:
      type: object
      required:
        - totalApprovedBudget
        - totalForecastBudget
        - totalActualSpend
        - totalRemaining
        - activeProjects
        - totalResources
        - openDemands
        - overallocatedCount
      properties:
        totalApprovedBudget: { type: number }
        totalForecastBudget: { type: number }
        totalActualSpend: { type: number }
        totalRemaining: { type: number }
        activeProjects: { type: integer }
        totalResources: { type: integer }
        openDemands: { type: integer }
        overallocatedCount: { type: integer }
    BudgetVsActualPoint:
      type: object
      required: [year, month, approved, forecast, actual]
      properties:
        year: { type: integer }
        month: { type: integer }
        approved: { type: number }
        forecast: { type: number }
        actual: { type: number }
    UtilizationRow:
      type: object
      required: [resourceId, resourceName, year, month, allocationPct, capacityPct]
      properties:
        resourceId: { type: integer }
        resourceName: { type: string }
        year: { type: integer }
        month: { type: integer }
        allocationPct: { type: number }
        capacityPct: { type: number }
...
    DemandVsCapacityRow:
      type: object
      required: [skill, demandFte, capacityFte, gap]
      properties:
        skill: { type: string }
        demandFte: { type: number }
        capacityFte: { type: number }
        gap: { type: number }
...
    OpexCapexRow:
      type: object
      required: [year, month, opex, capex]
      properties:
        year: { type: integer }
        month: { type: integer }
        opex: { type: number }
        capex: { type: number }
    SkillShortageRow:
      type: object
      required: [skill, year, month, demandFte, capacityFte, shortageFte]
      properties:
        skill: { type: string }
        year: { type: integer }
        month: { type: integer }
        demandFte: { type: number }
        capacityFte: { type: number }
        shortageFte: { type: number }
...